Rename GenesisValidatorHandler to GenesisValidator
