+++ casper/src/test/scala/coop/rchain/casper/util/comm/CasperPacketHandlerSpec.scala
@@ -130,73 +130,6 @@ class CasperPacketHandlerSpec extends WordSpec {
   }
 
   "CasperPacketHandler" when {
-    "in GenesisValidator state" should {
-
-      "respond on UnapprovedBlock messages with BlockApproval" in {
-        implicit val ctx = Scheduler.global
-        val fixture      = setup()
-        import fixture._
-
-        implicit val engineCell: EngineCell[Task] =
-          Cell.unsafe[Task, Engine[Task]](Engine.noop)
-        val packetHandler     = new CasperPacketHandler[Task]
-        val expectedCandidate = ApprovedBlockCandidate(Some(genesis), requiredSigs)
-
-        val unapprovedBlock  = BlockApproverProtocolTest.createUnapproved(requiredSigs, genesis)
-        val unapprovedPacket = BlockApproverProtocolTest.unapprovedToPacket(unapprovedBlock)
-        val test = for {
-          _ <- engineCell.set(
-                new GenesisValidatorHandler(runtimeManager, validatorId, shardId, bap)
-              )
-          _             <- packetHandler.handle(local).apply(unapprovedPacket)
-          blockApproval = BlockApproverProtocol.getBlockApproval(expectedCandidate, validatorId)
-          expectedPacket = ProtocolHelper.packet(
-            local,
-            networkId,
-            transport.BlockApproval,
-            blockApproval.toByteString
-          )
-          _ = {
-            val lastMessage = transportLayer.requests.last
-            assert(lastMessage.peer == local && lastMessage.msg == expectedPacket)
-          }
-        } yield ()
-        test.unsafeRunSync
-      }
-
-      "should not respond to any other message" in {
-        implicit val ctx = Scheduler.global
-        val fixture      = setup()
-        import fixture._
-
-        implicit val engineCell: EngineCell[Task] =
-          Cell.unsafe[Task, Engine[Task]](Engine.noop)
-        val packetHandler = new CasperPacketHandler[Task]
-
-        val approvedBlockRequest = ApprovedBlockRequest("test")
-        val packet1              = Packet(transport.ApprovedBlockRequest.id, approvedBlockRequest.toByteString)
-        val test = for {
-          _ <- engineCell.set(
-                new GenesisValidatorHandler(runtimeManager, validatorId, shardId, bap)
-              )
-          _    <- packetHandler.handle(local)(packet1)
-          head = transportLayer.requests.head
-          response = packet(
-            local,
-            networkId,
-            transport.NoApprovedBlockAvailable,
-            NoApprovedBlockAvailable(approvedBlockRequest.identifier, local.toString).toByteString
-          )
-          _            = assert(head.peer == local && head.msg == response)
-          _            = transportLayer.reset()
-          blockRequest = BlockRequest("base16Hash", ByteString.copyFromUtf8("base16Hash"))
-          packet2      = Packet(transport.BlockRequest.id, blockRequest.toByteString)
-          _            <- packetHandler.handle(local)(packet2)
-          _            = assert(transportLayer.requests.isEmpty)
-        } yield ()
-        test.unsafeRunSync
-      }
-    }
 
     "in  BootstrapCasperHandler state" should {
       "make a transition to ApprovedBlockReceivedHandler once ApprovedBlock has been received" in {
